backend/requirements.txt
# Django Backend Requirements
# Core Django and database dependencies

Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
mongoengine>=0.29.1
python-decouple>=3.8
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.2.2

# Development dependencies
django-debug-toolbar>=4.1.0